requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio==4.44.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# HTTP clients
requests==2.32.4
httpx==0.25.2

# Data processing
pandas==2.0.3
numpy==1.24.4

# ML/AI (optional, for future LLM integration)
torch==2.1.1
transformers==4.46.3
accelerate==1.0.1

# System monitoring
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pathlib2==2.3.7
dataclasses-json==0.6.7